scripts/requirements.txt
progressbar33==2.4
PyYAML==6.0.2
Requests==2.32.4
rich==14.0.0
urllib3==1.26.19
XlsxWriter==3.2.0
